refactor: переписал присвоение классов комп-м при использовании classnames. Скрыл функциональность переключения темы из комп-а changeThemesBtn
