+++ movies_app/src/components/Footer/Footer.module.css
@@ -7,7 +7,7 @@
   box-shadow: 10px 10px 10px 15px #dbd2d2;
 }
 
-.footer__ligth {
+.footer__light {
   background: linear-gradient(#d3abb7, #d9d9d9);
   box-shadow: 10px 10px 10px 15px #d56363;
 }
