+++ movies_app/src/components/Header/Header.jsx
@@ -7,16 +7,16 @@ import { HeaderLink } from "../HeaderLink/HeaderLink";
 import { useTheme } from '../../hooks/useTheme';
 
 export const Header = () => {
-  const { isLigth } = useTheme();
+  const { isLight } = useTheme();
 
   return (
-    <header className={isLigth ? cn(s.header, s.header__ligth) : s.header}>
+    <header className={ cn(s.header, {[s.header__light]: isLight}) }>
       <div className={s.wrapper}>
         <img className={s.logo} src={logo} alt='logo'/>
         <nav className={s.nav}> 
           <HeaderLink to='/' pagesName='Home' />
           <HeaderLink to='/search' pagesName='Search' />
-          <HeaderLink to='/favorite' pagesName='Favotite' />
+          <HeaderLink to='/favorite' pagesName='Favorite' />
           <div className={s.signBlock}>
             <HeaderLink to='/signin' pagesName='Sign In' />
           </div>
