+++ movies_app/src/components/Header/Header.module.css
@@ -7,7 +7,7 @@
   position: relative;
 }
 
-.header__ligth {
+.header__light {
   background: linear-gradient(#d3abb7, #d9d9d9);
   box-shadow: 0px 2px 10px 5px #f9bfbf;
 }
