+++ movies_app/src/components/containers/MainPage/MainPage.module.css
@@ -20,7 +20,7 @@
   border-right: 1px solid #6c6c6c;
 }
 
-.title__ligth {
+.title__light {
   background: linear-gradient(#f5f5f5, rgb(223 223 223 / 40%));
   border-right: 1px solid #c78181;
   color: #000000;
