+++ movies_app/src/providers/ThemeProvider.jsx
@@ -3,10 +3,12 @@ import React, {useState} from "react";
 export const ContextTheme = React.createContext();
 
 export const ThemeProvider = ({children}) => {
-  const [isLigth, changeTheme] = useState(false);
+  const [isLight, changeTheme] = useState(false);
+
+  const toggleTheme = () => changeTheme(!isLight);
 
   return (
-    <ContextTheme.Provider value={{isLigth, changeTheme}}>
+    <ContextTheme.Provider value={{isLight, toggleTheme}}>
       {children}
     </ContextTheme.Provider>
   );
