+++ movies_app/src/utilities/setActiveLink.js
@@ -1,4 +1,4 @@
 import cn from 'classnames';
 import s from '../components/HeaderLink/HeaderLink.module.css';
 
-export const setActiveLink = ({ isActive }) => isActive ? cn(s.link, s.active) : s.link;
+export const setActiveLink = ({ isActive }) => cn(s.link, {[s.active]: isActive});
